requirements.txt
Flask==2.2.2
PyMuPDF==1.25.2
Pillow==11.3.0
pytesseract==0.3.9
sentence-transformers==2.2.2
python-docx==0.8.11
